reafactor: move errors into separate module
