+++ rust/examples/read_delta_table.rs
@@ -1,5 +1,5 @@
 #[tokio::main(flavor = "current_thread")]
-async fn main() -> Result<(), deltalake::DeltaTableError> {
+async fn main() -> Result<(), deltalake::errors::DeltaTableError> {
     let table_path = "./tests/data/delta-0.8.0";
     let table = deltalake::open_table(table_path).await?;
     println!("{table}");
