+++ rust/src/lib.rs
@@ -87,6 +87,7 @@ pub mod builder;
 pub mod data_catalog;
 pub mod delta;
 pub mod delta_config;
+pub mod errors;
 pub mod operations;
 pub mod partitions;
 pub mod schema;
@@ -97,8 +98,6 @@ pub mod time_utils;
 #[cfg(all(feature = "arrow"))]
 pub mod table_state_arrow;
 
-#[cfg(all(feature = "arrow", feature = "parquet"))]
-pub mod checkpoints;
 #[cfg(all(feature = "arrow", feature = "parquet"))]
 pub mod delta_arrow;
 #[cfg(feature = "datafusion")]
@@ -116,6 +115,8 @@ pub use object_store::{path::Path, Error as ObjectStoreError, ObjectMeta, Object
 pub use operations::DeltaOps;
 
 // convenience exports for consumers to avoid aligning crate versions
+#[cfg(all(feature = "arrow", feature = "parquet"))]
+pub use action::checkpoints;
 #[cfg(feature = "arrow")]
 pub use arrow;
 #[cfg(feature = "datafusion")]
