+++ rust/src/operations/transaction/conflict_checker.rs
@@ -7,8 +7,9 @@ use object_store::ObjectStore;
 use super::CommitInfo;
 use crate::action::{Action, Add, DeltaOperation, MetaData, Protocol, Remove};
 use crate::delta_config::IsolationLevel;
+use crate::errors::{DeltaResult, DeltaTableError};
 use crate::storage::commit_uri_from_version;
-use crate::{table_state::DeltaTableState, DeltaResult, DeltaTableError};
+use crate::table_state::DeltaTableState;
 
 #[cfg(feature = "datafusion")]
 use super::state::AddContainer;
