+++ rust/src/storage/utils.rs
@@ -1,15 +1,16 @@
 //! Utility functions for working across Delta tables
 
 use std::collections::HashMap;
+use std::sync::Arc;
 
-use crate::action::Add;
-use crate::builder::DeltaTableBuilder;
-use crate::{DeltaResult, DeltaTableError};
 use chrono::{DateTime, NaiveDateTime, Utc};
 use futures::{StreamExt, TryStreamExt};
 use object_store::path::Path;
 use object_store::{DynObjectStore, ObjectMeta, Result as ObjectStoreResult};
-use std::sync::Arc;
+
+use crate::action::Add;
+use crate::builder::DeltaTableBuilder;
+use crate::errors::{DeltaResult, DeltaTableError};
 
 /// Copies the contents from the `from` location into the `to` location
 pub async fn copy_table(
