+++ rust/tests/command_optimize.rs
@@ -1,5 +1,9 @@
 #![cfg(all(feature = "arrow", feature = "parquet"))]
 
+use std::time::SystemTime;
+use std::time::UNIX_EPOCH;
+use std::{collections::HashMap, error::Error, sync::Arc};
+
 use arrow::datatypes::Schema as ArrowSchema;
 use arrow::{
     array::{Int32Array, StringArray},
@@ -7,16 +11,14 @@ use arrow::{
     record_batch::RecordBatch,
 };
 use deltalake::action::{Action, Remove};
+use deltalake::errors::DeltaTableError;
 use deltalake::operations::optimize::{create_merge_plan, MetricDetails, Metrics};
 use deltalake::operations::DeltaOps;
 use deltalake::writer::{DeltaWriter, RecordBatchWriter};
-use deltalake::{DeltaTable, DeltaTableError, PartitionFilter, SchemaDataType, SchemaField};
+use deltalake::{DeltaTable, PartitionFilter, SchemaDataType, SchemaField};
 use parquet::file::properties::WriterProperties;
 use rand::prelude::*;
 use serde_json::json;
-use std::time::SystemTime;
-use std::time::UNIX_EPOCH;
-use std::{collections::HashMap, error::Error, sync::Arc};
 use tempdir::TempDir;
 
 struct Context {
